Only allow to remove your own contacts

+++ src/main/java/ui/controller/RemoveContact.java
@@ -1,10 +1,11 @@
 package ui.controller;
 
-import domain.model.Role;
-import domain.model.Utility;
+import domain.db.DbException;
+import domain.model.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 public class RemoveContact extends RequestHandler {
     @Override
@@ -12,13 +13,19 @@ public class RemoveContact extends RequestHandler {
         Role[] authRoles = {Role.ADMIN, Role.USER};
         Utility.checkRole(request, authRoles);
         String destination = "RedirectController?command=ContactsPage";
+        HttpSession session = request.getSession();
+        Person person = (Person) session.getAttribute("person");
         String confirmation = request.getParameter("confirmation");
 
         if (!confirmation.isEmpty() && confirmation.equals("Remove")) {
-            int contactid = Integer.parseInt(request.getParameter("contactid"));
-            getService().removeOneContact(contactid);
-            //request.setAttribute("contactRemovedMessage", "Contact has been removed");
-            destination = "RedirectController?command=RemoveContactSucces";
+            try {
+                int contactid = Integer.parseInt(request.getParameter("contactid"));
+                Contact contact = getService().getOneContact(contactid);
+                if (!contact.getUserid().equals(person.getUserid())) return destination;
+                getService().removeOneContact(contactid);
+                //request.setAttribute("contactRemovedMessage", "Contact has been removed");
+                destination = "RedirectController?command=RemoveContactSucces";
+            } catch (NumberFormatException | DbException ignored) {}
         }
         return destination;
     }
